# openapi.yaml
openapi: 3.1.0
info:
  title: FireRedASR2S REST API
  description: |
    工业级语音识别服务，基于 FastAPI 构建。提供高效的 ASR（自动语音识别）、
    VAD（语音活动检测）、LID（语言识别）、标点恢复等功能，采用异步处理以提高性能。
  version: 1.0.0
  contact:
    name: FireRedASR2S
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
  - url: /
    description: 相对路径（配合网关/代理使用）
...
paths:
...
  /api/v1/system/transcribe:
    post:
      tags: [system]
      summary: 一站式语音识别
      description: 支持 ASR、VAD、LID、标点预测的完整流程
      operationId: systemTranscribe
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                  description: 音频文件
                uttid:
                  type: string
                  description: 话语ID
                enable_vad:
                  type: boolean
                  default: true
                  description: 启用VAD
                enable_lid:
                  type: boolean
                  default: true
                  description: 启用LID
                enable_punc:
                  type: boolean
                  default: true
                  description: 启用标点
                asr_type:
                  type: string
                  default: aed
                  description: ASR类型
                return_timestamp:
                  type: boolean
                  default: false
                  description: 返回时间戳
      responses:
        '200':
          description: 识别成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscribeSuccessResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/system/transcribe/submit:
    post:
      tags: [system]
      summary: 提交异步转录任务
      description: |
        适用于长时间音频（如会议录音）。立即返回 job_id，客户端轮询：
        - GET /system/transcribe/status/{job_id} 获取任务状态
        - GET /system/transcribe/result/{job_id} 获取识别结果
      operationId: systemTranscribeSubmit
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                  description: 音频文件
                uttid:
                  type: string
                  description: 话语ID
                enable_vad:
                  type: boolean
                  default: true
                enable_lid:
                  type: boolean
                  default: true
                enable_punc:
                  type: boolean
                  default: true
                asr_type:
                  type: string
                  default: aed
                return_timestamp:
                  type: boolean
                  default: false
      responses:
        '200':
          description: 任务已提交
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      job_id:
                        type: string
                        format: uuid
        '500':
          description: 服务器错误
  /api/v1/system/transcribe/status/{job_id}:
    get:
      tags: [system]
      summary: 查询转录任务状态
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功（data.status 为 pending/processing/completed/failed；任务不存在时 code=4005）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      job_id:
                        type: string
                      status:
                        type: string
                        enum: [pending, processing, completed, failed]
                      filename:
                        type: string
                      created_at:
                        type: number
  /api/v1/system/transcribe/result/{job_id}:
    get:
      tags: [system]
      summary: 获取转录结果
      description: 仅当 status=completed 时返回完整识别结果；status=failed 时返回错误信息
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功（completed 时 data 为识别结果；failed 时返回错误；进行中时提示稍后重试；任务不存在时 code=4005）
...
components:
  schemas:
    SuccessResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 状态码，0 表示成功
          example: 0
        message:
          type: string
          description: 响应消息
          example: success
        data:
          description: 响应数据，结构因接口而异
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 错误码
          example: 5000
        message:
          type: string
          description: 错误描述
        details:
          description: 错误详情
...
    TranscribeSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                uttid:
                  type: string
                  description: 话语ID
                text:
                  type: string
                  description: 识别文本
                dur_s:
                  type: number
                  description: 音频时长（秒）
                sentences:
                  type: array
                  items: {}
                  description: 分句结果
                vad_segments_ms:
                  type: array
                  items: {}
                  description: VAD 分段（毫秒）
                words:
                  type: array
                  items: {}
                  description: 词级时间戳（return_timestamp 为 true 时返回）
                processing_time_ms:
                  type: integer
                  description: 处理耗时（毫秒）
...
  securitySchemes: {}
...